refactor: CSV 関連モジュール名を整理する
asset_balance の CSV 解析を本体へ統合し、csv_parse を csv_util へ改名して参照を更新。\n\nCo-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
